relax content security policy to allow unsafe-eval in production
The vue instances were not getting initialized in production without this.  Removing unsafe-eval from the development policy helped me confirm that this was the culprit and you see a warning message about it that cannot be seen in production.  In production, it simply refused to render the content (vue instance) and instead, replaced the vue div with an html comment.

Reference:
https://stackoverflow.com/questions/50951409/creating-vue-instance-comments-out-everything